+++ Server/src/controllers/GradeCapacityController.js
@@ -1,6 +1,8 @@
 const { Sequelize, DataTypes } = require('sequelize');
 const sequelize = require('../config/seq');
 const NotificationController = require('./NotificationController');
+const UsersController = require('./UsersController');
+const { verifyPreschool } = require('../config/token_validation');
 
 const Grade = require('../models/grade_capacity')(sequelize, DataTypes);
 const Preschool = require('../models/preschool')(sequelize, DataTypes);
@@ -46,50 +48,15 @@ const GradesController = {
         }
     },
 
-    async addGradeCapacity(req, res) {
-        const { grade, capacity } = req.body;
-        try {
-            if (grade && capacity) {
-                const newGrade = await Grade.create({ grade: grade, capacity: capacity });
-                return res.status(201).json({
-                    message: 'Grade Capacity added successfully',
-                    Grade: newGrade,
-                });
-            }
-            else {
-                return res.status(500).json({ message: "Incomplete information. Please provide both grade and capacity." });
-            }
-        } catch (error) {
-            return res.status(500).json({ message: error.message });
-        }
-    },
-
-    async trackWaitlist(preschool, grade) {
-        const available = await GradesController.checkGradeCapacity(preschool, grade);
-        //if it has space, pick the oldest waitlisted application, change status and notify parent
-        if (available) {
-            const waitlisted = await Application.findOne({
-                where: {
-                    status: 'Waitlist',
-                },
-                order: [['createdAt', 'ASC']],
-                include: User
-            });
-            if (waitlisted) {
-                waitlisted.status = "Pending"
-                waitlisted.save();
-
-                //notify parent 
-                if (waitlisted.User.role_name == "Parent") {
-                    await NotificationController.pushSingleNotification(waitlisted.User.email, "Update: The Wait Is Over!", "Your child's application is now under review. Please book an appointment for the evaluation to proceed.");
-                }
-            }
-        }
-    },
-
     async updateGradeCapacities(req, res) {
         const { gradesList, preschool_id } = req.body;
         try {
+
+            // access control 
+            if (await UsersController.getCurrentUserRole(req) == "Admin" && await verifyPreschool(preschool_id, req) == false) {
+                return res.status(403).json({ message: "Access Denied! You're Unauthorized To Perform This Action." });
+            }
+
             //remove existing 
             await Grade.destroy({
                 where: { preschool_id: preschool_id }
@@ -144,6 +111,31 @@ const GradesController = {
             throw error;
         }
     },
+
+
+    async trackWaitlist(preschool, grade) {
+        const available = await GradesController.checkGradeCapacity(preschool, grade);
+        //if it has space, pick the oldest waitlisted application, change status and notify parent
+        if (available) {
+            const waitlisted = await Application.findOne({
+                where: {
+                    status: 'Waitlist',
+                },
+                order: [['createdAt', 'ASC']],
+                include: User
+            });
+            if (waitlisted) {
+                waitlisted.status = "Pending"
+                waitlisted.save();
+
+                //notify parent 
+                if (waitlisted.User.role_name == "Parent") {
+                    await NotificationController.pushSingleNotification(waitlisted.User.email, "Update: The Wait Is Over!", "Your child's application is now under review. Please book an appointment for the evaluation to proceed.");
+                }
+            }
+        }
+    },
+
 };
